requirements.txt
numpy
python-telegram-bot
python-dotenv
oauthlib
requests
requests-oauthlib
python-dateutil
icalendar